move all tests to typogrify/tests
