+++ typogrify/templatetags/typogrify_tags.py
@@ -505,12 +505,3 @@ def typogrify(text):
     text = number_suffix(text)
 
     return text
-
-
-def _test():
-    import doctest
-    doctest.testmod()
-
-
-if __name__ == "__main__":
-    _test()
