+++ typogrify/tests/test_fuzzydate.py
@@ -7,7 +7,7 @@ from django.test import TestCase
 from typogrify.templatetags.typogrify_tags import fuzzydate
 
 
-class TestFuzzyDate(TestCase):
+class TestFuzzydate(TestCase):
 
     def setUp(self):
         settings.DATE_FORMAT = "F jS, Y"
