+++ typogrify/tests/test_tags.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+import doctest
+
+from typogrify.templatetags.typogrify_tags import *
+
+
+if __name__ == "__main__":
+    doctest.testmod()
